requirements.txt
Flask-RESTful
Flask-PyMongo
gunicorn
pymongo[gssapi,srv,tls]